Rename command types
