+++ Tests/NIOIMAPCoreTests/Grammar/CommandType/CommandType+Tests.swift
@@ -24,7 +24,7 @@ class CommandType_Tests: EncodeTestClass {
 extension CommandType_Tests {
 
     func testEncode() {
-        let inputs: [(NIOIMAP.CommandType, String, UInt)] = [
+        let inputs: [(NIOIMAP.Command, String, UInt)] = [
             (.list(nil, "", .mailbox(""), []), "LIST \"\" \"\" RETURN ()", #line),
             (.namespace, "NAMESPACE", #line)
         ]
